+++ geode-core/src/integrationTest/java/org/apache/geode/internal/io/LocatorJmxSerialFilterIntegrationTest.java
@@ -23,56 +23,60 @@ import static org.apache.geode.distributed.ConfigurationProperties.JMX_MANAGER;
 import static org.apache.geode.distributed.ConfigurationProperties.JMX_MANAGER_PORT;
 import static org.apache.geode.distributed.ConfigurationProperties.JMX_MANAGER_START;
 import static org.apache.geode.distributed.ConfigurationProperties.LOG_FILE;
-import static org.apache.geode.internal.AvailablePortHelper.getRandomAvailableTCPPort;
+import static org.apache.geode.internal.AvailablePortHelper.getRandomAvailableTCPPorts;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assumptions.assumeThat;
 
 import java.io.File;
-import java.io.IOException;
 
 import org.junit.Before;
-import org.junit.Ignore;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.contrib.java.lang.system.RestoreSystemProperties;
 import org.junit.rules.TemporaryFolder;
 
-import org.apache.geode.distributed.ServerLauncher;
-import org.apache.geode.internal.serialization.filter.SanctionedSerializablesFilterPattern;
+import org.apache.geode.distributed.LocatorLauncher;
+import org.apache.geode.distributed.internal.InternalLocator;
+import org.apache.geode.internal.cache.InternalCache;
+import org.apache.geode.internal.serialization.filter.OpenMBeanFilterPattern;
+import org.apache.geode.management.ManagementService;
+import org.apache.geode.management.internal.SystemManagementService;
 import org.apache.geode.test.junit.rules.CloseableReference;
 
-public class ServerJdkGlobalSerialFilterIntegrationTest {
+public class LocatorJmxSerialFilterIntegrationTest {
 
-  private static final String NAME = "server";
-  private static final String JDK_SERIAL_FILTER_PROPERTY = "jdk.serialFilter";
+  private static final String NAME = "locator";
+  private static final String JMX_SERIAL_FILTER_PROPERTY =
+      "jmx.remote.rmi.server.serial.filter.pattern";
 
   private File workingDirectory;
+  private int locatorPort;
   private int jmxPort;
-  private String sanctionedSerializablesFilterPattern;
+  private String openMBeanFilterPattern;
 
   @Rule
-  public CloseableReference<ServerLauncher> server = new CloseableReference<>();
+  public CloseableReference<LocatorLauncher> locator = new CloseableReference<>();
   @Rule
   public RestoreSystemProperties restoreSystemProperties = new RestoreSystemProperties();
   @Rule
   public TemporaryFolder temporaryFolder = new TemporaryFolder();
 
   @Before
-  public void setUp() throws IOException {
+  public void setUp() throws Exception {
     workingDirectory = temporaryFolder.newFolder(NAME);
-    jmxPort = getRandomAvailableTCPPort();
-    sanctionedSerializablesFilterPattern = new SanctionedSerializablesFilterPattern().pattern();
+    int[] ports = getRandomAvailableTCPPorts(2);
+    locatorPort = ports[0];
+    jmxPort = ports[1];
+    openMBeanFilterPattern = new OpenMBeanFilterPattern().pattern();
   }
 
   @Test
-  @Ignore
-  // jdk.serialFilter does not apply to ServerLauncher
-  public void configuresJdkSerialFilter_onJava9orGreater() {
+  public void configuresJmxSerialFilter_onJava9orGreater() {
     assumeThat(isJavaVersionAtLeast(JAVA_9)).isTrue();
 
-    server.set(new ServerLauncher.Builder()
+    locator.set(new LocatorLauncher.Builder()
         .setMemberName(NAME)
-        .setDisableDefaultServer(true)
+        .setPort(locatorPort)
         .setWorkingDirectory(workingDirectory.getAbsolutePath())
         .set(HTTP_SERVICE_PORT, "0")
         .set(JMX_MANAGER, "true")
@@ -83,22 +87,22 @@ public class ServerJdkGlobalSerialFilterIntegrationTest {
         .get()
         .start();
 
-    assertThat(System.getProperty(JDK_SERIAL_FILTER_PROPERTY))
-        .as(JDK_SERIAL_FILTER_PROPERTY)
-        .isEqualTo(sanctionedSerializablesFilterPattern);
+    assertThat(isJmxManagerStarted())
+        .isTrue();
+    assertThat(System.getProperty(JMX_SERIAL_FILTER_PROPERTY))
+        .as(JMX_SERIAL_FILTER_PROPERTY)
+        .isEqualTo(openMBeanFilterPattern);
   }
 
   @Test
-  @Ignore
-  // jdk.serialFilter does not apply to ServerLauncher
-  public void changesEmptyJdkSerialFilter_onJava9orGreater() {
+  public void changesEmptyJmxSerialFilter_onJava9orGreater() {
     assumeThat(isJavaVersionAtLeast(JAVA_9)).isTrue();
 
-    System.setProperty(JDK_SERIAL_FILTER_PROPERTY, "");
+    System.setProperty(JMX_SERIAL_FILTER_PROPERTY, "");
 
-    server.set(new ServerLauncher.Builder()
+    locator.set(new LocatorLauncher.Builder()
         .setMemberName(NAME)
-        .setDisableDefaultServer(true)
+        .setPort(locatorPort)
         .setWorkingDirectory(workingDirectory.getAbsolutePath())
         .set(HTTP_SERVICE_PORT, "0")
         .set(JMX_MANAGER, "true")
@@ -109,21 +113,23 @@ public class ServerJdkGlobalSerialFilterIntegrationTest {
         .get()
         .start();
 
-    assertThat(System.getProperty(JDK_SERIAL_FILTER_PROPERTY))
-        .as(JDK_SERIAL_FILTER_PROPERTY)
-        .isEqualTo(sanctionedSerializablesFilterPattern);
+    assertThat(isJmxManagerStarted())
+        .isTrue();
+    assertThat(System.getProperty(JMX_SERIAL_FILTER_PROPERTY))
+        .as(JMX_SERIAL_FILTER_PROPERTY)
+        .isEqualTo(openMBeanFilterPattern);
   }
 
   @Test
-  public void doesNotChangeNonEmptySerialFilter_onJava9orGreater() {
+  public void doesNotChangeNonEmptyJmxSerialFilter_onJava9orGreater() {
     assumeThat(isJavaVersionAtLeast(JAVA_9)).isTrue();
 
-    String existingSerialFilter = "!*";
-    System.setProperty(JDK_SERIAL_FILTER_PROPERTY, existingSerialFilter);
+    String existingJmxSerialFilter = "!*";
+    System.setProperty(JMX_SERIAL_FILTER_PROPERTY, existingJmxSerialFilter);
 
-    server.set(new ServerLauncher.Builder()
+    locator.set(new LocatorLauncher.Builder()
         .setMemberName(NAME)
-        .setDisableDefaultServer(true)
+        .setPort(locatorPort)
         .setWorkingDirectory(workingDirectory.getAbsolutePath())
         .set(HTTP_SERVICE_PORT, "0")
         .set(JMX_MANAGER, "true")
@@ -134,18 +140,20 @@ public class ServerJdkGlobalSerialFilterIntegrationTest {
         .get()
         .start();
 
-    assertThat(System.getProperty(JDK_SERIAL_FILTER_PROPERTY))
-        .as(JDK_SERIAL_FILTER_PROPERTY)
-        .isEqualTo(existingSerialFilter);
+    assertThat(isJmxManagerStarted())
+        .isTrue();
+    assertThat(System.getProperty(JMX_SERIAL_FILTER_PROPERTY))
+        .as(JMX_SERIAL_FILTER_PROPERTY)
+        .isEqualTo(existingJmxSerialFilter);
   }
 
   @Test
-  public void doesNotConfigureJdkSerialFilter_onJava8() {
+  public void doesNotConfigureJmxSerialFilter_onJava8() {
     assumeThat(isJavaVersionAtMost(JAVA_1_8)).isTrue();
 
-    server.set(new ServerLauncher.Builder()
+    locator.set(new LocatorLauncher.Builder()
         .setMemberName(NAME)
-        .setDisableDefaultServer(true)
+        .setPort(locatorPort)
         .setWorkingDirectory(workingDirectory.getAbsolutePath())
         .set(HTTP_SERVICE_PORT, "0")
         .set(JMX_MANAGER, "true")
@@ -156,20 +164,22 @@ public class ServerJdkGlobalSerialFilterIntegrationTest {
         .get()
         .start();
 
-    assertThat(System.getProperty(JDK_SERIAL_FILTER_PROPERTY))
-        .as(JDK_SERIAL_FILTER_PROPERTY)
+    assertThat(isJmxManagerStarted())
+        .isTrue();
+    assertThat(System.getProperty(JMX_SERIAL_FILTER_PROPERTY))
+        .as(JMX_SERIAL_FILTER_PROPERTY)
         .isNull();
   }
 
   @Test
-  public void doesNotChangeEmptyJdkSerialFilter_onJava8() {
+  public void doesNotChangeEmptyJmxSerialFilter_onJava8() {
     assumeThat(isJavaVersionAtMost(JAVA_1_8)).isTrue();
 
-    System.setProperty(JDK_SERIAL_FILTER_PROPERTY, "");
+    System.setProperty(JMX_SERIAL_FILTER_PROPERTY, "");
 
-    server.set(new ServerLauncher.Builder()
+    locator.set(new LocatorLauncher.Builder()
         .setMemberName(NAME)
-        .setDisableDefaultServer(true)
+        .setPort(locatorPort)
         .setWorkingDirectory(workingDirectory.getAbsolutePath())
         .set(HTTP_SERVICE_PORT, "0")
         .set(JMX_MANAGER, "true")
@@ -180,21 +190,23 @@ public class ServerJdkGlobalSerialFilterIntegrationTest {
         .get()
         .start();
 
-    assertThat(System.getProperty(JDK_SERIAL_FILTER_PROPERTY))
-        .as(JDK_SERIAL_FILTER_PROPERTY)
+    assertThat(isJmxManagerStarted())
+        .isTrue();
+    assertThat(System.getProperty(JMX_SERIAL_FILTER_PROPERTY))
+        .as(JMX_SERIAL_FILTER_PROPERTY)
         .isEqualTo("");
   }
 
   @Test
-  public void doesNotChangeNonEmptyJdkSerialFilter_onJava8() {
+  public void doesNotChangeNonEmptyJmxSerialFilter_onJava8() {
     assumeThat(isJavaVersionAtMost(JAVA_1_8)).isTrue();
 
-    String existingSerialFilter = "!*";
-    System.setProperty(JDK_SERIAL_FILTER_PROPERTY, existingSerialFilter);
+    String existingJmxSerialFilter = "!*";
+    System.setProperty(JMX_SERIAL_FILTER_PROPERTY, existingJmxSerialFilter);
 
-    server.set(new ServerLauncher.Builder()
+    locator.set(new LocatorLauncher.Builder()
         .setMemberName(NAME)
-        .setDisableDefaultServer(true)
+        .setPort(locatorPort)
         .setWorkingDirectory(workingDirectory.getAbsolutePath())
         .set(HTTP_SERVICE_PORT, "0")
         .set(JMX_MANAGER, "true")
@@ -205,8 +217,20 @@ public class ServerJdkGlobalSerialFilterIntegrationTest {
         .get()
         .start();
 
-    assertThat(System.getProperty(JDK_SERIAL_FILTER_PROPERTY))
-        .as(JDK_SERIAL_FILTER_PROPERTY)
-        .isEqualTo(existingSerialFilter);
+    assertThat(isJmxManagerStarted())
+        .isTrue();
+    assertThat(System.getProperty(JMX_SERIAL_FILTER_PROPERTY))
+        .as(JMX_SERIAL_FILTER_PROPERTY)
+        .isEqualTo(existingJmxSerialFilter);
+  }
+
+  private SystemManagementService getSystemManagementService() {
+    InternalLocator locator = (InternalLocator) this.locator.get().getLocator();
+    InternalCache cache = locator.getCache();
+    return (SystemManagementService) ManagementService.getManagementService(cache);
+  }
+
+  private boolean isJmxManagerStarted() {
+    return getSystemManagementService().isManager();
   }
 }
