+++ geode-core/src/main/java/org/apache/geode/internal/InternalDataSerializer.java
@@ -368,6 +368,7 @@ public abstract class InternalDataSerializer extends DataSerializer {
       String filterPattern = new SanctionedSerializablesFilterPattern()
           .append(distributionConfig.getSerializableObjectFilter())
           .pattern();
+      System.out.println("JC debug: filterPattern=" + filterPattern);
       serializationFilter = new DelegatingObjectInputFilterFactory()
           .create(filterPattern, loadSanctionedClassNames(services));
     } else {
