+++ gradle.properties
@@ -67,8 +67,8 @@ compileJVM=
 compileJVMVer=8
 
 #JAVA_HOME to be used by tests
-testJVM=/usr/local/Cellar/openjdk@11/11.0.12/libexec/openjdk.jdk/Contents/Home
-testJVMVer=11
+testJVM=
+testJVMVer=8
 
 repeat = 100
